drivers: mtk-connectivity: Checkout to Samsung R source

+++ drivers/misc/mediatek/connectivity/common/common_main/core/wmt_ic_6628.c
@@ -783,7 +783,7 @@ static INT32 mt6628_sw_init(P_WMT_HIF_CONF pWmtHifConf)
 	ctrlPa1 = 0;
 	ctrlPa2 = 0;
 	wmt_core_ctrl(WMT_CTRL_GET_PATCH_NUM, &ctrlPa1, &ctrlPa2);
-	patch_num = ctrlPa1;
+	patch_num = (UINT32)ctrlPa1;
 	WMT_INFO_FUNC("patch total num = [%d]\n", patch_num);
 
 	/* 9.3 Multi-patch Patch download */
